apply style ;D
