+++ NewsApp/NewsApp/Realm/DataSources/NewsArticleDataSource.swift
@@ -14,34 +14,6 @@ protocol NewsArticleDataSourceDelegate: AnyObject {
     func newsArticleDataSourceDelegate(didUpdateArticles dataSource: NewsArticleDataSource)
 }
 
-/// Use this data source if you are displaying the data in a UITableView.
-/// Set `tableView` property to reference the UITableView that is presenting the content.
-class NewsArticleTableViewDataSource: NewsArticleDataSource {
-    weak var tableView: UITableView?
-    
-    override func setObserver(forArticles articles: Results<ArticleDB>) -> NotificationToken? {
-        guard let tableView = tableView else {
-            return nil
-        }
-        
-        return RealmHelper.observeRealmTableResults(tableView: tableView, results: articles)
-    }
-}
-
-/// Use this data source if you are displaying the data in a UICollectionView.
-/// Set `collectionView` property to reference the UICollectionView that is presenting the content.
-class NewsArticleCollectionViewDataSource: NewsArticleDataSource {
-    weak var collectionView: UICollectionView?
-    
-    override func setObserver(forArticles articles: Results<ArticleDB>) -> NotificationToken? {
-        guard let collectionView = collectionView else {
-            return nil
-        }
-        
-        return RealmHelper.observeRealmCollectionResults(collectionView: collectionView, results: articles)
-    }
-}
-
 class NewsArticleDataSource {
     private var articlesDb: Results<ArticleDB>?
     private var token: NotificationToken?
